requirements.txt
beautifulsoup4==4.12.3
fastapi==0.111.1
google-generativeai==0.8.3
httpx==0.27.0
jinja2==3.1.4
playwright==1.48.0
pydantic==2.8.2
pydantic-settings==2.4.0
pytest==8.3.2
uvicorn[standard]==0.30.6
